RF: move utility functions
